requirements.txt
httpx~=0.28.1
pydantic~=2.10.6
fastapi~=0.115.8
python-dotenv~=1.0.1
pytest~=8.3.4
pytest-asyncio~=0.25.3
uvicorn~=0.34.0
loguru~=0.7.2
aiohttp>=3.8.0
